A changed version of hos_side.cpp - the boos thread_goup gives a segmentation fault on CSC.
It does not appear to be a library thing (tried different libraries with
"-mt" flag also). The error appears in threads.join_all() after a mutex
destructor: ~shared_mutex().

Since only one device is used in the voxelization process, the boost
group threads is used if more devices are being initialized. This avoids
the problem but does not solve it.